ScrollableChat migration to TS
