+++ frontend/src/utils/getUserInfo.ts
@@ -22,28 +22,6 @@ export const isSameSender = (
   );
 };
 
-export const isSameSenderMargin = (
-  messages: Message[],
-  m: Message,
-  i: number,
-  userId: string
-) => {
-  if (
-    i < messages.length - 1 &&
-    messages[i + 1].sender._id === m.sender._id &&
-    messages[i].sender._id !== userId
-  )
-    return 33;
-  else if (
-    (i < messages.length - 1 &&
-      messages[i + 1].sender._id !== m.sender._id &&
-      messages[i].sender._id !== userId) ||
-    (i === messages.length - 1 && messages[i].sender._id !== userId)
-  )
-    return 0;
-  else return 'auto';
-};
-
 export const isSameUser = (messages: Message[], m: Message, i: number) => {
   return i > 0 && messages[i - 1].sender._id === m.sender._id;
 };
